feat: add new ldes client source based on update in the ldes client
