+++ package.json
@@ -29,13 +29,14 @@
   "dependencies": {
     "@actions/core": "^1.5.0",
     "@actions/exec": "^1.1.0",
-    "@treecg/actor-init-ldes-client": "2.3.5",
+    "@treecg/actor-init-ldes-client": "2.3.7",
     "date-fns": "^2.23.0",
     "n3": "^1.11.1"
   },
   "devDependencies": {
     "@vercel/ncc": "^0.29.2",
     "cross-env": "^7.0.3",
+    "env-cmd": "^10.1.0",
     "prettier": "^2.3.2",
     "typescript": "^4.3.5"
   }
